Regulate heartbeats per-client rather than per-connection, fixes #72
The heartbeat fiber is started once in the client constructor instead of
separately in each HELLO handler, fixing #72 itself. This could cause
instability during reconnections, hence an instance variable
`@send_heartbeats` was introduced that is set at HELLO and unset at
disconnection. The heartbeat handler only sends a packet if this
variable is set; if not it simply waits for the next turn.